Navbar components styled
I have made the nav components float to the right by using span as  the coding was done in bootstrap. I've also changed the components colors by linking the html with the css which was not linked before. I have attached in the screenshot for better review of my work. Hope you like it !😊🤗