Navbar components style added
Fixes the issue #49 .

I have made the navbar components float to the right by using span as the coding was done in bootstrap. I've also changed the components colors by linking the html with the css which was not linked before. I have attached in the screenshot for better review of my work. Upon hovering the nav components it shows the hover color. I've also changed the logo color and its font-size . Hope you like it !😊🤗